Second Commit - About Section done
Done with the about section directly under the hero which contains the logo and "what we do" stuff. Nesting a link within a button is problematic HTML which I might change later but for now I'm too lazy. Similarily considering replacing the logo with the actual embedded youtube video we're supposed to have included.