Update About.css
The relevant CSS for the about section. I also included a very very bare attempt for the responsiveness bit (specifically making the burger menu with just html and css, might use java instead) which I did yesterday but forgot to commit. Heavily considering leaving the responsive bits for this specific page for the very last as I can see it taking alot of time which might risk that I miss an entire page or so if I spend too much time on the about page.