CSS Apply Final
I've essentially had to re-do the code from scratch. The header was getting super buggy so I dropped it and decided I'll insert it when I'm done with the application form, so that entire part of the code is gone now. Somehow the responsiveness seems to have figured itself out here so I'm satisfied with this for now. Will however return to fix some issues regarding requirement and add some more media-querries in the future.