Final Draft
As a consequence of losing my mind trying to figure out how to make this work,  and having to start from scratch 20 times over, I forgot to gradually update my commits here so this'll basically be the finished product dumped in one go sadly (if needed, you can ask me about anything regarding this code whenever). I'll properly adjust the header and add a footer later, and work on responsiveness towards the end - but this is basically the finished product besides those aspects (I hope, atleast).