Final Touches
This specific website is just weird. When I use inspect and choose the dimensions of an iphone screen, there's overflowing space on the right (a bit) and the it doesn't look THAT good. But when I simply shrink my page to match the iphone dimensions, or use any other method besides inspect, it looks perfectly fine. Not sure what to make of that, but I hope your inspect is better than mine when you view this for responsiveness.