Responsiveness + Menu
When it comes to this page, as opposed to changing the initial css to minimize the amount of media-querries I need to use, I decided to instead focus on the media-querries for the most part while leaving the initial css as it is (since I feel that I haven't really displayed whether or not I am competent at using media-querries or not in the other pages).  Added the menu too and changed the logo pic for a video.

Again, the inspect-tool where I choose an iphone and it auto-adjusts just doesn't work for me. Instead, I worked off how it looked like with restore-down + manually changing the width of the page from the outside. So if it does look funky when you use inspect, that's probably why and hence why I'd recommend changing the width manually instead.