Completion
Realized it is easier for me to just directly copy-paste the VS code entirely as opposed to checking for specific details and make sure everything is matching. I know it looks like I just made an entirely new web-page but the codes are pretty much the same expect very minor differences which is why I completely copy-pasted it.